Thompson/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
numpy>=1.21.0,<2.0.0
pydantic>=1.8.0,<2.0.0
orjson>=3.6.0,<4.0.0
PyYAML
numpy-financial
